functions/importing/requirements.txt
scipy
biopython
numpy
dssp
ipympl
matplotlib